move exchange func to wallet.go initialize exchange call on send page load split switch code to multi func
